BackEnd Working

+++ FontesBank/FontesBank/Models/UserModel.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace FontesBank.Models
 {
@@ -11,7 +8,7 @@ namespace FontesBank.Models
         public string Name { get; set; }
         public string Email { get; set; }
         public DateTime BirthDate { get; set; }
-        public int CurrentBalance { get; set; }
+        public float CurrentBalance { get; set; }
         public string Address { get; set; }
 
     }
